--- src/main/java/com/main_39/Spring/order/Order.java
@@ -1,40 +1,0 @@
-package com.main_39.Spring.order;
-
-import com.main_39.Spring.audit.Auditable;
-import com.main_39.Spring.member.entity.Kakao;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@Table(name = "orders")
-public class Order extends Auditable {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long orderId;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "kakao_id")
-    private Kakao kakao;
-
-    @OneToMany(mappedBy = "order")
-    private List<OrderMenu> orderMenus = new ArrayList<>();
-
-    private int count;
-}
