--- src/main/java/com/main_39/Spring/order/OrderMenuPostRequestDto.java
@@ -1,9 +1,0 @@
-package com.main_39.Spring.order;
-
-import lombok.Getter;
-
-@Getter
-public class OrderMenuPostRequestDto {
-    private Long menuId;
-    private int count;
-}
