--- src/main/java/com/main_39/Spring/order/OrderPostRequestDto.java
@@ -1,16 +1,0 @@
-package com.main_39.Spring.order;
-
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-public class OrderPostRequestDto {
-    private long kakao_id;
-
-    private List<OrderMenuPostRequestDto> orderMenus;
-
-    /**
-     * TODO: 구매자 ID 가지고 오기
-     */
-}
