--- src/main/java/com/main_39/Spring/order/OrderRepository.java
@@ -1,6 +1,0 @@
-package com.main_39.Spring.order;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-}
